requirements.txt
# core
streamlit==1.48.0
python-dotenv==1.1.1
openai==1.99.3
wikipedia==1.4.0
SQLAlchemy>=2.0
PyMySQL>=1.1
supabase>=2.6
mutagen==1.47.0

# TTS
google-cloud-texttospeech==2.27.0